First 14 posts (up to, not including DART post) with static images
And BTW, these are not counting the img tag at the top ... those still use the assets. We could change that to avoid using the baseurl, etc., or just keep one image for each post that we reserve for that image. Either way will be more scalable than what we have now, and we're not going to get 1000 posts, this blog won't last that long, I'm sure.